WebApi/requirements.txt
Cerberus==0.9.2
cffi==1.3.0
Django==1.8.5
django-filter==0.12.0
djangorestframework==3.3.2
Events==0.2.1
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
Markdown==2.6.5
MarkupSafe==0.23
pycparser==2.14
pymongo==2.9.2
python-magic==0.4.10
requests==2.8.1
six==1.10.0
Werkzeug==0.10.4
wheel==0.24.0
